Support of CtScanner based queries
- Extraction of code to CtBaseQuery to make code cleaner and to allow
reuse of queries
- CtFunction can return Iterator now too
