+++ src/main/java/spoon/reflect/visitor/chain/CtQueryImpl.java
@@ -16,53 +16,35 @@
  */
 package spoon.reflect.visitor.chain;
 
-import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import spoon.Launcher;
-import spoon.SpoonException;
 import spoon.reflect.declaration.CtElement;
-import spoon.reflect.visitor.CtScanner;
 import spoon.reflect.visitor.Filter;
 
 /**
- * Contains the default implementation of the generic {@link CtQuery} methods
+ * The facade of {@link CtBaseQuery} which represents a query bound to the {@link CtElement},
+ * which is the constant input of this query.
+ * It is used by {@link CtElement} implementations of {@link CtQueryable}.
+ *
+ * @param <O> - the type of element which is produced by the last step of the query
  */
 public class CtQueryImpl<O> implements CtQuery<O> {
 
 	/**
-	 * All the constant inputs of this query. It can be null
+	 * All the constant inputs of this query.
 	 */
 	private List<Object> inputs;
 
 	/**
-	 * first step of the query
+	 * The {@link CtBaseQuery} which provides low level query behavior for this query.
 	 */
-	private Step firstStep;
-	/**
-	 * last step of the query
-	 */
-	private Step lastStep;
-	/**
-	 * the Step which wraps final Consumer, which is always present in each (even empty) query
-	 */
-	private Step tail;
-
-	private boolean logging = false;
-	private QueryFailurePolicy failurePolicy = QueryFailurePolicy.FAIL;
-
-	public CtQueryImpl() {
-		tail = new TailConsumer();
-		tail.setName("TailConsumer");
-		firstStep = tail;
-		lastStep = tail;
-	}
+	private final CtBaseQueryImpl<Object> query;
 
 	@SuppressWarnings("unchecked")
 	public <T> CtQueryImpl(T input) {
-		this();
+		query = new CtBaseQueryImpl<>();
 		setInput((O) input);
 	}
 
@@ -100,49 +82,32 @@ public class CtQueryImpl<O> implements CtQuery<O> {
 
 	@SuppressWarnings("unchecked")
 	@Override
+	public <R> CtQuery<R> map(CtLazyFunction<?, R> queryStep) {
+		query.map(queryStep);
+		return (CtQuery<R>) this;
+	}
+
+	@SuppressWarnings("unchecked")
+	@Override
 	public <I, R> CtQuery<R> map(CtFunction<I, R> function) {
-		add(new FunctionWrapper(function));
+		query.map(function);
 		return (CtQuery<R>) this;
 	}
 
 	@SuppressWarnings("unchecked")
 	@Override
 	public <T extends CtElement> CtQuery<T> filterChildren(Filter<T> filter) {
-		add(new ChildrenFilteringWrapper(filter));
+		query.filterChildren(filter);
 		return (CtQuery<T>) this;
 	}
 
-
-	/**
-	 * Sends the input parameter as input of the whole query
-	 * through chain of query steps and calls output.accept(element)
-	 * for each element produced by this query.
-	 *
-	 * @param input
-	 * @param output
-	 */
 	@SuppressWarnings("unchecked")
-	private void forEach(CtConsumer<O> output, Object input) {
-		tail.setNext((CtConsumer<Object>) output);
-		try {
-			if (input == null) {
-				if (inputs != null) {
-					for (Object in : inputs) {
-						firstStep.accept(in);
-					}
-				}
-			} else {
-				if (inputs != null) {
-					throw new SpoonException("Do not add QueryStep inputs if you want to use query for extra input");
-				}
-				firstStep.accept(input);
-			}
-		} finally {
-			tail.setNext(null);
+	public <R> void forEach(CtConsumer<R> consumer) {
+		for (Object input : inputs) {
+			query.apply(input, (CtConsumer<Object>) consumer);
 		}
 	}
 
-	@Override
 	public List<O> list() {
 		final List<O> list = new ArrayList<>();
 		forEach(new CtConsumer<O>() {
@@ -150,29 +115,20 @@ public class CtQueryImpl<O> implements CtQuery<O> {
 			public void accept(O out) {
 				list.add(out);
 			}
-		}, null);
+		});
 		return list;
 	}
 
 	@Override
 	public CtQuery<O> name(String name) {
-		if (lastStep == tail) {
-			throw new SpoonException("Cannot set name of the step on the chain with no step");
-		}
-		lastStep.setName(name);
-		return this;
-	}
-	@Override
-	public CtQuery<O> failurePolicy(QueryFailurePolicy policy) {
-		if (lastStep == tail) {
-			throw new SpoonException("Cannot set ignoreIncompatibleInput of the step on the chain with no step");
-		}
-		failurePolicy = policy;
+		query.name(name);
 		return this;
 	}
 
-	private boolean isLogging() {
-		return logging;
+	@Override
+	public CtQuery<O> failurePolicy(QueryFailurePolicy policy) {
+		query.failurePolicy(policy);
+		return this;
 	}
 
 	/**
@@ -183,270 +139,7 @@ public class CtQueryImpl<O> implements CtQuery<O> {
 	 * Reason: Query chains are used internally during writing of log messages too. So it would write logs for ever...
 	 */
 	public CtQuery<O> logging(boolean logging) {
-		this.logging = logging;
+		query.logging(logging);
 		return this;
-	}
-
-	private void add(Step step) {
-		if (lastStep == tail) {
-			firstStep = step;
-			lastStep = step;
-		} else {
-			lastStep.setNext(step);
-			lastStep = step;
-		}
-		step.setNext(tail);
-		name(String.valueOf(getLength()));
-	}
-
-	/**
-	 * @return number of steps of this query
-	 */
-	public int getLength() {
-		int len = 0;
-		Step s = firstStep;
-		while (s != tail) {
-			len++;
-			s = (Step) s.getNext();
-		}
-		return len;
-	}
-
-	/**
-	 * Is used to log that invocation was not processed
-	 * @param step the step which thrown CCE
-	 * @param e
-	 * @param parameters
-	 */
-	private void onClassCastException(Step step, ClassCastException e, Object... parameters) {
-		if (step.isFailOnCCE()) {
-			throw e;
-		} else if (Launcher.LOGGER.isTraceEnabled()) {
-			//log expected CCE ... there might be some unexpected too!
-			Launcher.LOGGER.trace(e);
-		}
-		log(step, e.getMessage(), parameters);
-	}
-
-	private void log(Step step, String message, Object... parameters) {
-		if (isLogging() && Launcher.LOGGER.isInfoEnabled()) {
-			StringBuilder sb = new StringBuilder("Step ");
-			sb.append(step.getName()).append(") ");
-			sb.append(message);
-			for (int i = 0; i < parameters.length; i++) {
-				sb.append("\nParameter ").append(i + 1).append(") ");
-				sb.append(parameters[i]);
-			}
-			Launcher.LOGGER.info(sb.toString());
-		}
-	}
-
-	/**
-	 * Each Step must implement it
-	 */
-	private interface Step extends CtConsumer<Object> {
-		/**
-		 * @return next {@link CtConsumer}
-		 */
-		CtConsumer<Object> getNext();
-		/**
-		 * sets next {@link CtConsumer}
-		 */
-		void setNext(CtConsumer<Object> next);
-		/**
-		 * @return name of this Step - for debugging purposes
-		 */
-		String getName();
-		/**
-		 * @param name of the step - for debugging purposes
-		 */
-		void setName(String name);
-		/**
-		 * @return true if this step should throw {@link ClassCastException} in case of
-		 * step input type incompatibility
-		 */
-		boolean isFailOnCCE();
-	}
-
-	/**
-	 * abstract step which knows next Consumer and then name
-	 */
-	private abstract class AbstractStep implements Step {
-		private String name;
-		protected CtConsumer<Object> next;
-		@Override
-		public CtConsumer<Object> getNext() {
-			return next;
-		}
-		@Override
-		public void setNext(CtConsumer<Object> next) {
-			this.next = next;
-		}
-		@Override
-		public String getName() {
-			return name;
-		}
-		@Override
-		public void setName(String name) {
-			this.name = name;
-		}
-		@Override
-		public boolean isFailOnCCE() {
-			return failurePolicy == QueryFailurePolicy.FAIL;
-		}
-		public final void accept(Object input) {
-			if (input == null) {
-				return;
-			}
-			log(this, "received", input);
-			processInput(input);
-		}
-		protected abstract void processInput(Object input);
-	}
-
-	/**
-	 * There is always one TailConsumer in each query, which sends result to output
-	 */
-	private class TailConsumer extends AbstractStep {
-		@Override
-		public void processInput(Object out) {
-			if (next != null) {
-				try {
-					next.accept(out);
-				} catch (ClassCastException e) {
-					onClassCastException(this, e, out);
-				}
-			}
-		}
-	}
-
-	/**
-	 * a step which calls Function. Implements contract of {@link CtQuery#map(CtFunction)}
-	 */
-	private class FunctionWrapper extends AbstractStep {
-		private CtFunction<Object, Object> fnc;
-
-		@SuppressWarnings("unchecked")
-		FunctionWrapper(CtFunction<?, ?> code) {
-			fnc = (CtFunction<Object, Object>) code;
-		}
-
-		@SuppressWarnings("unchecked")
-		@Override
-		public void processInput(Object input) {
-			Object result;
-			try {
-				result = fnc.apply(input);
-			} catch (ClassCastException e) {
-				onClassCastException(this, e, input);
-				return;
-			}
-			if (result == null) {
-				return;
-			}
-			if (result instanceof Boolean) {
-				//the code is a predicate. send the input to output if result is true
-				if ((Boolean) result) {
-					next.accept(input);
-				} else {
-					log(this, "Skipped element, because CtFunction#accept(input) returned false", input);
-				}
-				return;
-			}
-			if (result instanceof Iterable) {
-				//send each item of Iterable to the next step
-				for (Object out : (Iterable<Object>) result) {
-					next.accept(out);
-				}
-				return;
-			}
-			if (result.getClass().isArray()) {
-				//send each item of Array to the next step
-				for (int i = 0; i < Array.getLength(result); i++) {
-					next.accept(Array.get(result, i));
-				}
-				return;
-			}
-			next.accept(result);
-		}
-	}
-
-	/**
-	 * a step which scans all children of input element and only elements matching filter go to the next step
-	 */
-	private class ChildrenFilteringWrapper extends CtScanner implements Step {
-
-		private String name;
-		protected CtConsumer<Object> next;
-		private Filter<CtElement> filter;
-
-		@SuppressWarnings("unchecked")
-		ChildrenFilteringWrapper(Filter<? extends CtElement> filter) {
-			this.filter = (Filter<CtElement>) filter;
-		}
-
-		@Override
-		public void accept(Object input) {
-			if (input == null) {
-				return;
-			}
-			log(this, "CtScanner received", input);
-			scan(input);
-		}
-		@Override
-		public void scan(CtElement element) {
-			processFilter(element);
-			super.scan(element);
-		}
-
-		private void processFilter(CtElement element) {
-			if (element == null) {
-				return;
-			}
-			boolean matches = false;
-			try {
-				matches = filter.matches(element);
-			} catch (ClassCastException e) {
-				onClassCastException(this, e, element);
-				return;
-			}
-			if (matches) {
-				//send input to output, because Fitler.matches returned true
-				next.accept(element);
-			} else {
-				log(this, "Skipped child element, because Filter#matches(input) returned false", element);
-			}
-		}
-
-		@Override
-		public CtConsumer<Object> getNext() {
-			return next;
-		}
-		@Override
-		public void setNext(CtConsumer<Object> next) {
-			this.next = next;
-		}
-		@Override
-		public String getName() {
-			return name;
-		}
-		@Override
-		public void setName(String name) {
-			this.name = name;
-		}
-		@Override
-		public boolean isFailOnCCE() {
-			//never fail on CCE during Filter.matching
-			return false;
-		}
-	}
-
-	/**
-	 * Interface used to query and manipulate elements in a functional manner.
-	 *
-	 * @param &lt;T> - the type of accepted elements
-	 */
-	private interface CtConsumer<T> {
-		void accept(T t);
 	}
 }
