+++ src/main/java/spoon/support/reflect/declaration/CtElementImpl.java
@@ -31,8 +31,9 @@ import spoon.reflect.visitor.DefaultJavaPrettyPrinter;
 import spoon.reflect.visitor.Filter;
 import spoon.reflect.visitor.ModelConsistencyChecker;
 import spoon.reflect.visitor.Query;
-import spoon.reflect.visitor.chain.CtFunction;
 import spoon.reflect.visitor.chain.CtQueryImpl;
+import spoon.reflect.visitor.chain.CtFunction;
+import spoon.reflect.visitor.chain.CtLazyFunction;
 import spoon.reflect.visitor.chain.CtQuery;
 import spoon.reflect.visitor.filter.AnnotationFilter;
 import spoon.support.util.EmptyClearableList;
@@ -255,6 +256,11 @@ public abstract class CtElementImpl implements CtElement, Serializable {
 
 	public <E extends CtElement> List<E> getElements(Filter<E> filter) {
 		return filterChildren(filter).list();
+	}
+
+	@Override
+	public <T> CtQuery<T> map(CtLazyFunction<?, T> queryStep) {
+		return new CtQueryImpl<T>(this).map(queryStep);
 	}
 
 	@Override
